Load old List::Util XS component if needed for Scalar::Util or Sub::Util
Old versions of Scalar::Util and Sub::Util rely on List::Util to provide
their functions, by loading the XS component of List::Util.  We can't
depend on the new versions that have their own functions, because this
dist will be moving to being pure perl in the future.

We first need to detect if the old versions or Scalar::Util or Sub::Util
are being used.  If so, we load the old XS component, using the old
version number.  While doing this, we "localize" the stashes for the
newer versions of the modules, including our own, so that only the
modules that need the old code are modified.